KA-137 - Move checking methods to utils, adjust tests
