+++ onboarding-task/.idea/codeStyleSettings.xml
@@ -39,7 +39,9 @@
           <option name="RIGHT_MARGIN" value="72" />
         </MarkdownNavigatorCodeStyleSettings>
         <TypeScriptCodeStyleSettings>
+          <option name="FORCE_SEMICOLON_STYLE" value="true" />
           <option name="USE_DOUBLE_QUOTES" value="false" />
+          <option name="FORCE_QUOTE_STYlE" value="true" />
           <option name="OBJECT_LITERAL_WRAP" value="0" />
           <option name="SPACES_WITHIN_IMPORTS" value="true" />
         </TypeScriptCodeStyleSettings>
@@ -116,6 +118,7 @@
           </indentOptions>
         </codeStyleSettings>
         <codeStyleSettings language="TypeScript">
+          <option name="RIGHT_MARGIN" value="100" />
           <option name="LINE_COMMENT_ADD_SPACE" value="false" />
           <indentOptions>
             <option name="INDENT_SIZE" value="2" />
