+++ onboarding-task/src/utils/checkStatus.ts
