+++ onboarding-task/src/utils/isInsertEmpty.ts
